# compress/flate: read data with partial flush