… src/compress/flate/flate_test.go
… src/compress/flate/testdata/partial-block